+++ model/matriculaModel.js
@@ -21,7 +21,7 @@ class MatriculaModel {
 
     async postMatricula(codAluno, codDisciplina, semestre) {
         return new Promise((resolve, reject) => {
-            db.query('INSERT INTO matriculas (codAluno, codDisciplina, semestre, notas) VALUES (?,?,?)', [codAluno, codDisciplina, semestre, notas], (err, result) => {
+            db.query('INSERT INTO matriculas (codigoAluno, codigoDisciplina, semestre) VALUES (?,?,?)', [codAluno, codDisciplina, semestre], (err, result) => {
                 if(err) throw err;
                 resolve(result);
             })
@@ -30,7 +30,7 @@ class MatriculaModel {
 
     async updateMatricula(id, modification) {
         return new Promise((resolve, reject) => {
-            db.query('UPDATE matriculas SET ? WHERE id =?', [modification, id], (err, result) => {
+            db.query(`UPDATE matriculas SET ${modification} WHERE id = ${id}`, (err, result) => {
                 if(err) throw err;
                 resolve(result);
             })
@@ -38,13 +38,23 @@ class MatriculaModel {
     }
 
     async deleteMatricula(id) {
-        console.log('passei aqui')
-        return new Promise((resolve, reject) => {
-            db.query('DELETE FROM matriculas WHERE id =?', [id], (err, result) => {
-                if(err) throw err;
-                resolve(result);
+        const notas = new Promise((resolve, reject) =>{
+            db.query(`SELECT * FROM notas WHERE codigoMatricula = ${id}`,(err, result) => {
+            if (err) reject(err);
+            resolve(result);
             })
-        })
+        });
+        
+        if(JSON.stringify(await notas)== '[]'){
+            return new Promise((resolve, reject) => {
+                db.query('DELETE FROM matriculas WHERE id =?', [id], (err, result) => {
+                    if(err) throw err;
+                    resolve(result);
+                })
+            })
+        }else{
+            return "ErrorNotas";
+        }
     }
 }
